refactor : 커스텀 훅 사용하여 queries 분리
